Add files via upload
Files associated with the program, including the executable. The documentation is lackluster at the moment. I will update it at another time. Anyone that downloads this for whatever purposes that they desire it for should know that an antivirus may not like the program. Not sure why, I'm not far along enough at this time to know.